# docs/swagger-bundle.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Contacts App",
    "description": "Documentation for Contacts API",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "tags": [
    { "name": "Contacts", "description": "Operations with user contacts" },
    { "name": "Auth", "description": "Authentication and user management" }
  ],
  "servers": [
    { "url": "http://localhost:3000" },
    { "url": "https://nodejs-hw-mongodb-mw37.onrender.com" }
  ],
  "paths": {
    "/contacts": {
      "get": {
        "tags": ["Contacts"],
        "summary": "Get all contacts",
        "operationId": "getAllContacts",
        "description": "Retrieve a list of all contacts. Supports pagination and filtering.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "query", "name": "page", "schema": { "type": "integer", "example": 1 } },
          { "in": "query", "name": "perPage", "schema": { "type": "integer", "example": 10 } },
          { "in": "query", "name": "sortBy", "schema": { "type": "string", "example": "name" } },
          { "in": "query", "name": "sortOrder", "schema": { "type": "string", "enum": ["asc", "desc"] } },
          { "in": "query", "name": "isFavourite", "schema": { "type": "boolean" } }
        ],
        "responses": {
          "200": {
            "description": "Successfully fetched contacts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "integer", "example": 200 },
                    "message": { "type": "string", "example": "Success" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/contact" }
                        },
                        "page": { "type": "integer", "example": 1 },
                        "perPage": { "type": "integer", "example": 10 },
                        "totalItems": { "type": "integer", "example": 50 },
                        "totalPages": { "type": "integer", "example": 5 },
                        "hasNextPage": { "type": "boolean", "example": true },
                        "hasPreviousPage": { "type": "boolean", "example": false }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/404" }
        }
      },
      "post": {
        "tags": ["Contacts"],
        "summary": "Create a new contact",
        "operationId": "createContact",
        "description": "Create a new contact and return the created object",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "John Doe" },
                  "phoneNumber": { "type": "string", "example": "+380991112233" },
                  "email": { "type": "string", "example": "john.doe@example.com" },
                  "isFavourite": { "type": "boolean", "example": true },
                  "contactType": { "type": "string", "enum": ["work", "home", "personal"], "example": "work" },
                  "photo": { "type": "string", "format": "binary" }
                },
                "required": ["name", "phoneNumber", "contactType"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contact created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "integer", "example": 201 },
                    "message": { "type": "string", "example": "Successfully created a contact!" },
                    "data": { "$ref": "#/components/schemas/contact" }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid input data" },
          "404": { "$ref": "#/components/responses/404" }
        }
      }
    },
    "/contacts/{id}": {
      "get": {
        "tags": ["Contacts"],
        "summary": "Get contact by ID",
        "operationId": "getContactById",
        "description": "Retrieve a contact by its unique ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Contact found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "integer", "example": 200 },
                    "data": { "$ref": "#/components/schemas/contact" }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/404" }
        }
      },
      "patch": {
        "tags": ["Contacts"],
        "summary": "Update contact",
        "operationId": "updateContact",
        "description": "Update an existing contact by ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "John Doe" },
                  "phoneNumber": { "type": "string", "example": "+380991112233" },
                  "email": { "type": "string", "example": "john.doe@example.com" },
                  "isFavourite": { "type": "boolean", "example": true },
                  "contactType": { "type": "string", "enum": ["work", "home", "personal"], "example": "work" },
                  "photo": { "type": "string", "format": "binary" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contact updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "integer", "example": 200 },
                    "message": { "type": "string", "example": "Contact updated successfully" },
                    "data": { "$ref": "#/components/schemas/contact" }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/404" }
        }
      },
      "delete": {
        "tags": ["Contacts"],
        "summary": "Delete contact",
        "operationId": "deleteContact",
        "description": "Delete a contact by its ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "204": {
            "description": "Contact deleted successfully"
          },
          "404": { "$ref": "#/components/responses/404" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer" }
    },
    "schemas": {
      "contact": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "example": "John Doe" },
          "phoneNumber": { "type": "string", "example": "+380991112233" },
          "email": { "type": "string", "example": "john.doe@example.com" },
          "isFavourite": { "type": "boolean", "example": true },
          "contactType": { "type": "string", "enum": ["work", "home", "personal"], "example": "work" },
          "photo": { "type": "string", "example": "https://example.com/photos/john.jpg" }
        }
      }
    },
    "responses": {
      "404": {
        "description": "Not Found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": { "type": "integer", "example": 404 },
                "message": { "type": "string", "example": "Contact not found" }
              }
            }
          }
        }
      }
    }
  }
}